Change destination field to GrinboxAddress type
