+++ src/common/crypto.rs
@@ -5,10 +5,6 @@ use std::fmt::Write;
 use super::base58::{ToBase58, FromBase58};
 use common::{ErrorKind, Result};
 use sha2::{Sha256, Digest};
-use rand::Rng;
-use rand::thread_rng;
-use ring::aead;
-use ring::{digest, pbkdf2};
 
 pub const GRINBOX_ADDRESS_VERSION_MAINNET: [u8; 2] = [1, 11];
 pub const GRINBOX_ADDRESS_VERSION_TESTNET: [u8; 2] = [1, 120];
@@ -116,74 +112,6 @@ pub fn verify_signature(challenge: &str, signature: &Signature, public_key: &Pub
     let secp = Secp256k1::new();
     secp.verify(&message, signature, public_key).map_err(|_| ErrorKind::Secp.into())
 }
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct EncryptedMessage {
-    #[serde(default)]
-    pub destination: Option<String>,
-    encrypted_message: String,
-    salt: String,
-    nonce: String,
-}
-
-impl EncryptedMessage {
-    pub fn new(message: String, destination: String, receiver_public_key: &PublicKey, secret_key: &SecretKey) -> Result<EncryptedMessage> {
-        let secp = Secp256k1::new();
-        let mut common_secret = receiver_public_key.clone();
-        common_secret.mul_assign(&secp, secret_key).map_err(|_| ErrorKind::Encryption)?;
-        let common_secret_ser = common_secret.serialize_vec(&secp, true);
-        let common_secret_slice = &common_secret_ser[1..33];
-
-        let salt: [u8; 8] = thread_rng().gen();
-        let nonce: [u8; 12] = thread_rng().gen();
-        let mut key = [0; 32];
-        pbkdf2::derive(&digest::SHA512, 100, &salt, common_secret_slice, &mut key);
-        let mut enc_bytes = message.as_bytes().to_vec();
-        let suffix_len = aead::CHACHA20_POLY1305.tag_len();
-        for _ in 0..suffix_len {
-            enc_bytes.push(0);
-        }
-        let sealing_key = aead::SealingKey::new(&aead::CHACHA20_POLY1305, &key)
-            .map_err(|_| ErrorKind::Encryption)?;
-        aead::seal_in_place(&sealing_key, &nonce, &[], &mut enc_bytes, suffix_len)
-            .map_err(|_| ErrorKind::Encryption)?;
-
-        Ok(EncryptedMessage {
-            destination: Some(destination),
-            encrypted_message: to_hex(enc_bytes),
-            salt: to_hex(salt.to_vec()),
-            nonce: to_hex(nonce.to_vec()),
-        })
-    }
-
-    pub fn key(&self, sender_public_key: &PublicKey, secret_key: &SecretKey) -> Result<[u8; 32]> {
-        let salt = from_hex(self.salt.clone()).map_err(|_| ErrorKind::Decryption)?;
-
-        let secp = Secp256k1::new();
-        let mut common_secret = sender_public_key.clone();
-        common_secret.mul_assign(&secp, secret_key).map_err(|_| ErrorKind::Decryption)?;
-        let common_secret_ser = common_secret.serialize_vec(&secp, true);
-        let common_secret_slice = &common_secret_ser[1..33];
-
-        let mut key = [0; 32];
-        pbkdf2::derive(&digest::SHA512, 100, &salt, common_secret_slice, &mut key);
-
-        Ok(key)
-    }
-
-    pub fn decrypt_with_key(&self, key: &[u8; 32]) -> Result<String> {
-        let mut encrypted_message = from_hex(self.encrypted_message.clone()).map_err(|_| ErrorKind::Decryption)?;
-        let nonce = from_hex(self.nonce.clone()).map_err(|_| ErrorKind::Decryption)?;
-
-        let opening_key = aead::OpeningKey::new(&aead::CHACHA20_POLY1305, key)
-            .map_err(|_| ErrorKind::Decryption)?;
-        let decrypted_data = aead::open_in_place(&opening_key, &nonce, &[], 0, &mut encrypted_message)
-            .map_err(|_| ErrorKind::Decryption)?;
-
-        String::from_utf8(decrypted_data.to_vec()).map_err(|_| ErrorKind::Decryption.into())
-    }
-}
-
 
 /// Encode the provided bytes into a hex string
 pub fn to_hex(bytes: Vec<u8>) -> String {
